fixes some dockerfiles, local option

+++ languages/elixirlang/run_benchmark.sh
@@ -1,17 +1,23 @@
 #!/bin/bash
 
-# Not working yet, can't make chromium work
-exit 0
+# Check if --local flag is provided
+if [[ "$1" == "--local" ]]; then
+    echo "Running Elixir benchmark locally..."
+    mix benchmark
+else
+    # Not working yet, can't make chromium work
+    exit 0
 
-IMAGE_NAME="elixir_benchmark"
-CONTAINER_NAME="elixir_benchmark_runner"
-OUTPUT_FILE="benchmarks.json"
-WORKDIR="/app"
+    IMAGE_NAME="elixir_benchmark"
+    CONTAINER_NAME="elixir_benchmark_runner"
+    OUTPUT_FILE="benchmarks.json"
+    WORKDIR="/app"
 
-docker build --platform=linux/arm64 -t $IMAGE_NAME .
+    docker build --platform=linux/arm64 -t $IMAGE_NAME .
 
-docker run --name $CONTAINER_NAME $IMAGE_NAME
+    docker run --name $CONTAINER_NAME $IMAGE_NAME
 
-docker cp $CONTAINER_NAME:$WORKDIR/$OUTPUT_FILE .
+    docker cp $CONTAINER_NAME:$WORKDIR/$OUTPUT_FILE .
 
-docker rm $CONTAINER_NAME
+    docker rm $CONTAINER_NAME
+fi
